backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
gunicorn==23.0.0
marshmallow-sqlalchemy==1.1.0
numpy==2.2.1
pytest==8.3.4
python-dotenv==1.0.1
python-magic==0.4.27